fix: Cast poll event bitwise OR results to short
- Cast OR'd event masks to `short` for `evPtr->events`, `taskPtr->events`
- Cast OR'd event masks to `short` when building `pfds[nfds].events`
- Cast `WaitFd()` poll `events | POLLPRI | POLLERR` assignment to `short`
- Avoid implicit integer promotion/sign conversion warnings and keep types
  consistent with `pollfd.events`/stored event fields

+++ nsproxy/nsproxylib.c
@@ -1588,7 +1588,7 @@ WaitFd(int fd, short events, long ms)
     int n;
 
     pfd.fd = fd;
-    pfd.events = events | POLLPRI | POLLERR;
+    pfd.events = (short)(events | POLLPRI | POLLERR);
     pfd.revents = pfd.events;
     do {
         n = ns_poll(&pfd, 1, ms);
